pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "design-research-analysis"
version = "0.1.0"
description = "Utilities for design research analysis workflows"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "The Design Research Collective" }]
maintainers = [{ name = "Christopher C. McComb", email = "ccm@cmu.edu" }]
keywords = ["analysis", "design-research", "hmm", "markov", "sequence"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = [
  "matplotlib>=3.8,<4",
  "numpy>=1.26,<3",
]

[project.optional-dependencies]
dev = [
  "build>=1.2,<2",
  "mypy>=1.10,<2",
  "pre-commit>=3.7,<5",
  "pytest>=8.2,<9",
  "pytest-cov>=7.0,<8",
  "ruff>=0.6.0,<1",
  "sphinx>=7.4,<9",
  "sphinx-rtd-theme>=2.0,<4",
  "twine>=5.1,<7",
  "uv>=0.6,<1",
]
table = []
seq = [
  "hmmlearn>=0.3.2,<1",
  "networkx>=3.3,<4",
  "scipy>=1.11,<2",
]
embeddings = ["sentence-transformers>=3.0,<4"]
lang = [
  "scikit-learn>=1.5,<2",
]
dimred = [
  "scikit-learn>=1.5,<2",
  "umap-learn>=0.5.7,<1",
]
stats = [
  "scipy>=1.11,<2",
  "statsmodels>=0.14,<1",
  "pandas>=2.2,<3",
]
all = [
  "hmmlearn>=0.3.2,<1",
  "networkx>=3.3,<4",
  "scipy>=1.11,<2",
  "sentence-transformers>=3.0,<4",
  "scikit-learn>=1.5,<2",
  "umap-learn>=0.5.7,<1",
  "statsmodels>=0.14,<1",
  "pandas>=2.2,<3",
]

[project.scripts]
design-research-analysis = "design_research_analysis.cli:main"

[project.urls]
Homepage = "https://github.com/cmudrc/design-research-analysis"
Repository = "https://github.com/cmudrc/design-research-analysis"
Issues = "https://github.com/cmudrc/design-research-analysis/issues"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
design_research_analysis = ["py.typed"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests", "examples", "scripts"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF", "D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D"]

[tool.mypy]
packages = ["design_research_analysis"]
python_version = "3.12"
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
  "hmmlearn",
  "hmmlearn.*",
  "matplotlib",
  "matplotlib.*",
  "networkx",
  "networkx.*",
  "sentence_transformers",
  "sentence_transformers.*",
  "sklearn",
  "sklearn.*",
  "umap",
  "umap.*",
  "pandas",
  "pandas.*",
  "statsmodels",
  "statsmodels.*",
]
ignore_missing_imports = true
